[WIP] Change: Move Type type to ir module
