+++ src/ir/mod.rs
@@ -1,5 +1,7 @@
 pub mod eir;
 pub mod printer;
+pub mod type_;
 
 pub use self::eir::{BindingKind, Constant, Identifier, Value};
 pub use self::printer::Printer;
+pub use self::type_::Type;
