+++ src/transform/type_infer/mod.rs
@@ -5,5 +5,4 @@ pub mod type_;
 pub mod type_env;
 pub mod type_infer;
 
-pub use self::type_::Type;
 pub use self::type_infer::TypeInfer;
